Put vertices on the same layer as the parent poly(line|mesh|face)
Previously, the layer for vertices defaulted to layer 0. Only
Polyline.add_vertex() allowed the layer to be specified. Presumably this
behavior was unintentional: while I can't find anything in the DXF
specification about this, it is typically considered an error to have
the vertices on a different layer. AutoCAD moves the vertices to the
parent layer (http://goo.gl/MflqsR), DraftSight does the same and
recognizes the issue during an error check (tested myself), and Layout
Editor puts the polylines on layer 0 (tested myself).

It is debatable whether the ability to put the vertices on a different
layer with Polyline.add_vertex() should be retained. It appears that
some less common programs may intentionally generate such DXFs
(http://goo.gl/p1hcfS). To remove this ability, simply delete the "if
not kwargs.get('layer')" test.